refactor : 바벨 플러그인 디렉토리 생성 & 분리된 runner 제거
기존 runner함수는 testManager 메서드로 동작할 수 있기 때문에 불필요한 분리를 제거.
바벨 플러그인을 디렉토리에 저장.
